+++ src/components/shop/product/detail/ProductDetailHeader.jsx
@@ -8,6 +8,7 @@ import instance from "../../../../utils/axios";
 
 import {toast} from "react-toastify";
 import DefaultButton from "../../../common/DefaultButton";
+import CartButton from "./CartButton";
 
 
 const ProductDetailHeader = ({data}) => {
@@ -38,91 +39,6 @@ const ProductDetailHeader = ({data}) => {
     useEffect(() => {
         console.log(xy)
     },[xy])
-
-    // 옵션을 담는 세션
-    const addOptions = () => {
-
-        let sessionOptions = {
-            itemId: data?.id,
-            options: data?.options
-        }
-
-        console.log("sessionOptions", sessionOptions);
-
-        let storageOpitons = localStorage.getItem("options");
-        storageOpitons = storageOpitons ? JSON.parse(storageOpitons) : [];
-
-        let validItemId = storageOpitons.find(options =>
-            sessionOptions.itemId === options.itemId)
-
-        if (!validItemId) {
-            storageOpitons.push(sessionOptions);
-            localStorage.setItem("options", JSON.stringify(storageOpitons));
-        }
-    }
-
-    // 비로그인 회원 장바구니 담기
-    const addCart = () => {
-        const itemList = session; // 추가하려는 상품 리스트
-
-        if (session.length > 0) {
-            // localStorage에서 cart 데이터를 가져오기
-            let storageCart = localStorage.getItem("cart");
-            storageCart = storageCart ? JSON.parse(storageCart) : []; // JSON 파싱 또는 빈 배열로 초기화
-        
-            // itemList를 순회하여 storageCart에 추가 또는 업데이트
-            itemList.forEach((item) => {
-                const existingItemIndex = storageCart.findIndex(
-                    (cartItem) =>
-                        cartItem.cartItemId === item.cartItemId && cartItem.option_name === item.option_name
-                );
-        
-                if (existingItemIndex !== -1) {
-                    // 동일한 itemId와 optionId가 있을 경우 count 업데이트
-                    storageCart[existingItemIndex].count += item.count;
-                } else {
-                    // 새로운 항목일 경우 추가
-                    storageCart.push(item);
-                }
-            });
-        
-            // 업데이트된 storageCart를 다시 localStorage에 저장
-            localStorage.setItem("cart", JSON.stringify(storageCart));
-            addOptions();
-
-            toast.success("장바구니에 담았습니다!");            
-        } else {
-            toast.error("옵션을 먼저 선택해주세요")
-        }
-    
-    };
-
-    // post 통신
-    const handlePostCart = () => {
-        const data = stocks;
-        let error = 0;
-
-        data.map((v, i) => {
-            instance({
-                url: "/cart/add",
-                method: "post",
-                data: v
-            }).then(res=>{
-                // console.log("성공했습니다 ",i)
-                // toast.success("장바구니에 담았습니다!");
-            }).catch(err=>{
-                console.log(err);
-                error ++
-                // toast.error("장바구니에 상품을 담는 중에 오류가 발생했습니다.")
-            })
-        })
-
-        if (error > 0) {
-            toast.error("장바구니에 상품을 담는 중에 오류가 발생했습니다.")
-        } else {
-            toast.success("장바구니에 담았습니다!")
-        }
-    }
 
     //선택옵션 추가핸들러
     const handleSelectChange = (_,val) => {
@@ -333,12 +249,7 @@ const ProductDetailHeader = ({data}) => {
                     {stocks[0] ? <span className="price">총 상품 금액 {priceCalculator()} 원</span>:<span></span>}
 
                     <div className={"buttons-container row"}>
-                    <DefaultButton className={"default long"} onClick={() => {
-                        addCart();
-                        handlePostCart();
-                    }}>
-                        장바구니
-                    </DefaultButton>
+                    <CartButton data={data} session={session} stocks={stocks}/>
                     <DefaultButton className={"primary long"} onClick={purchaseHandler}>구매하기</DefaultButton>
                     </div>
                 </div>
